@@ -1,27 +1,27 @@
 {
-  "_from": "@sinclair/typebox@^0.24.1",
-  "_id": "@sinclair/typebox@0.24.51",
+  "_from": "@sinclair/typebox@^0.25.16",
+  "_id": "@sinclair/typebox@0.25.23",
   "_inBundle": false,
-  "_integrity": "sha512-1P1OROm/rdubP5aFDSZQILU0vrLCJ4fvHt6EoqHEM+2D/G5MK3bIaymUKLit8Js9gbns5UyJnkP/TZROLw4tUA==",
+  "_integrity": "sha512-VEB8ygeP42CFLWyAJhN5OklpxUliqdNEUcXb4xZ/CINqtYGTjL5ukluKdKzQ0iWdUxyQ7B0539PAUhHKrCNWSQ==",
   "_location": "/@sinclair/typebox",
   "_phantomChildren": {},
   "_requested": {
     "type": "range",
     "registry": true,
-    "raw": "@sinclair/typebox@^0.24.1",
+    "raw": "@sinclair/typebox@^0.25.16",
     "name": "@sinclair/typebox",
     "escapedName": "@sinclair%2ftypebox",
     "scope": "@sinclair",
-    "rawSpec": "^0.24.1",
+    "rawSpec": "^0.25.16",
     "saveSpec": null,
-    "fetchSpec": "^0.24.1"
+    "fetchSpec": "^0.25.16"
   },
   "_requiredBy": [
     "/@jest/schemas"
   ],
-  "_resolved": "https://registry.npmjs.org/@sinclair/typebox/-/typebox-0.24.51.tgz",
-  "_shasum": "645f33fe4e02defe26f2f5c0410e1c094eac7f5f",
-  "_spec": "@sinclair/typebox@^0.24.1",
+  "_resolved": "https://registry.npmjs.org/@sinclair/typebox/-/typebox-0.25.23.tgz",
+  "_shasum": "1c15b0d2b872d89cc0f47c7243eacb447df8b8bd",
+  "_spec": "@sinclair/typebox@^0.25.16",
   "_where": "C:\\Users\\betsa\\ClientsideAPP\\AppReactWebCours\\node_modules\\@jest\\schemas",
   "author": {
     "name": "sinclairzx81"
@@ -36,13 +36,25 @@
     "@sinclair/hammer": "^0.17.1",
     "@types/chai": "^4.3.3",
     "@types/mocha": "^9.1.1",
-    "@types/node": "^18.7.13",
-    "ajv": "^8.11.0",
+    "@types/node": "^18.11.9",
+    "ajv": "^8.11.2",
     "ajv-formats": "^2.1.1",
     "chai": "^4.3.6",
     "mocha": "^9.2.2",
     "prettier": "^2.7.1",
-    "typescript": "^4.8.2"
+    "typescript": "^4.9.3"
+  },
+  "exports": {
+    "./compiler": "./compiler/index.js",
+    "./conditional": "./conditional/index.js",
+    "./custom": "./custom/index.js",
+    "./errors": "./errors/index.js",
+    "./format": "./format/index.js",
+    "./guard": "./guard/index.js",
+    "./hash": "./hash/index.js",
+    "./system": "./system/index.js",
+    "./value": "./value/index.js",
+    ".": "./typebox.js"
   },
   "homepage": "https://github.com/sinclairzx81/typebox#readme",
   "keywords": [
@@ -68,5 +80,5 @@
     "test": "hammer task test"
   },
   "types": "./typebox.d.ts",
-  "version": "0.24.51"
+  "version": "0.25.23"
 }
