@@ -1,8 +1,8 @@
 {
   "_from": "body-parser@^1.20.1",
-  "_id": "body-parser@1.20.1",
+  "_id": "body-parser@1.20.2",
   "_inBundle": false,
-  "_integrity": "sha512-jWi7abTbYwajOytWCQc37VulmWiRae5RyTpaCyDcS5/lMdtwSz5lOpDE67srw/HYe35f1z3fDQw+3txg7gNtWw==",
+  "_integrity": "sha512-ml9pReCu3M61kGlqoTm2umSXTlRTuGTx0bfYj+uIUKKYycG5NtSbeetV3faSU6R7ajOPw0g/J1PvK4qNy7s5bA==",
   "_location": "/body-parser",
   "_phantomChildren": {},
   "_requested": {
@@ -18,8 +18,8 @@
   "_requiredBy": [
     "/@expo/dev-server"
   ],
-  "_resolved": "https://registry.npmjs.org/body-parser/-/body-parser-1.20.1.tgz",
-  "_shasum": "b1812a8912c195cd371a3ee5e66faa2338a5c668",
+  "_resolved": "https://registry.npmjs.org/body-parser/-/body-parser-1.20.2.tgz",
+  "_shasum": "6feb0e21c4724d06de7ff38da36dad4f57a747fd",
   "_spec": "body-parser@^1.20.1",
   "_where": "C:\\Users\\betsa\\ClientsideAPP\\AppReactWebCours\\node_modules\\@expo\\dev-server",
   "bugs": {
@@ -39,7 +39,7 @@
   ],
   "dependencies": {
     "bytes": "3.1.2",
-    "content-type": "~1.0.4",
+    "content-type": "~1.0.5",
     "debug": "2.6.9",
     "depd": "2.0.0",
     "destroy": "1.2.0",
@@ -47,25 +47,25 @@
     "iconv-lite": "0.4.24",
     "on-finished": "2.4.1",
     "qs": "6.11.0",
-    "raw-body": "2.5.1",
+    "raw-body": "2.5.2",
     "type-is": "~1.6.18",
     "unpipe": "1.0.0"
   },
   "deprecated": false,
   "description": "Node.js body parsing middleware",
   "devDependencies": {
-    "eslint": "8.24.0",
+    "eslint": "8.34.0",
     "eslint-config-standard": "14.1.1",
-    "eslint-plugin-import": "2.26.0",
+    "eslint-plugin-import": "2.27.5",
     "eslint-plugin-markdown": "3.0.0",
     "eslint-plugin-node": "11.1.0",
-    "eslint-plugin-promise": "6.0.1",
+    "eslint-plugin-promise": "6.1.1",
     "eslint-plugin-standard": "4.1.0",
     "methods": "1.1.2",
-    "mocha": "10.0.0",
+    "mocha": "10.2.0",
     "nyc": "15.1.0",
     "safe-buffer": "5.2.1",
-    "supertest": "6.3.0"
+    "supertest": "6.3.3"
   },
   "engines": {
     "node": ">= 0.8",
@@ -91,5 +91,5 @@
     "test-ci": "nyc --reporter=lcov --reporter=text npm test",
     "test-cov": "nyc --reporter=html --reporter=text npm test"
   },
-  "version": "1.20.1"
+  "version": "1.20.2"
 }
