@@ -1,31 +1,31 @@
 {
-  "_from": "supports-color@^7.0.0",
-  "_id": "supports-color@7.2.0",
+  "_from": "supports-color@^8.0.0",
+  "_id": "supports-color@8.1.1",
   "_inBundle": false,
-  "_integrity": "sha512-qpCAvRl9stuOHveKsn7HncJRvv501qIacKzQlO/+Lwxc9+0q2wLyv4Dfvt80/DPn2pqOBsJdDiogXGR9+OvwRw==",
+  "_integrity": "sha512-MpUEN2OodtUzxvKQl72cUF7RQ5EiHsGvSsVG0ia9c5RbWGL2CI4C7EpPS8UTBIplnlzZiNuV56w+FuNxy3ty2Q==",
   "_location": "/jest-worker/supports-color",
   "_phantomChildren": {},
   "_requested": {
     "type": "range",
     "registry": true,
-    "raw": "supports-color@^7.0.0",
+    "raw": "supports-color@^8.0.0",
     "name": "supports-color",
     "escapedName": "supports-color",
-    "rawSpec": "^7.0.0",
+    "rawSpec": "^8.0.0",
     "saveSpec": null,
-    "fetchSpec": "^7.0.0"
+    "fetchSpec": "^8.0.0"
   },
   "_requiredBy": [
     "/jest-worker"
   ],
-  "_resolved": "https://registry.npmjs.org/supports-color/-/supports-color-7.2.0.tgz",
-  "_shasum": "1b7dcdcb32b8138801b3e478ba6a51caa89648da",
-  "_spec": "supports-color@^7.0.0",
+  "_resolved": "https://registry.npmjs.org/supports-color/-/supports-color-8.1.1.tgz",
+  "_shasum": "cd6fc17e28500cff56c1b86c0a7fd4a54a73005c",
+  "_spec": "supports-color@^8.0.0",
   "_where": "C:\\Users\\betsa\\ClientsideAPP\\AppReactWebCours\\node_modules\\jest-worker",
   "author": {
     "name": "Sindre Sorhus",
     "email": "sindresorhus@gmail.com",
-    "url": "sindresorhus.com"
+    "url": "https://sindresorhus.com"
   },
   "browser": "browser.js",
   "bugs": {
@@ -38,17 +38,22 @@
   "deprecated": false,
   "description": "Detect whether a terminal supports color",
   "devDependencies": {
-    "ava": "^1.4.1",
-    "import-fresh": "^3.0.0",
-    "xo": "^0.24.0"
+    "ava": "^2.4.0",
+    "import-fresh": "^3.2.2",
+    "xo": "^0.35.0"
   },
   "engines": {
-    "node": ">=8"
+    "node": ">=10"
+  },
+  "exports": {
+    "node": "./index.js",
+    "default": "./browser.js"
   },
   "files": [
     "index.js",
     "browser.js"
   ],
+  "funding": "https://github.com/chalk/supports-color?sponsor=1",
   "homepage": "https://github.com/chalk/supports-color#readme",
   "keywords": [
     "color",
@@ -81,5 +86,5 @@
   "scripts": {
     "test": "xo && ava"
   },
-  "version": "7.2.0"
+  "version": "8.1.1"
 }
