@@ -1,26 +1,26 @@
 {
-  "_from": "nocache@^2.1.0",
-  "_id": "nocache@2.1.0",
+  "_from": "nocache@^3.0.1",
+  "_id": "nocache@3.0.4",
   "_inBundle": false,
-  "_integrity": "sha512-0L9FvHG3nfnnmaEQPjT9xhfN4ISk0A8/2j4M37Np4mcDesJjHgEUfgPhdCyZuFI954tjokaIj/A3NdpFNdEh4Q==",
+  "_integrity": "sha512-WDD0bdg9mbq6F4mRxEYcPWwfA1vxd0mrvKOyxI7Xj/atfRHVeutzuWByG//jfm4uPzp0y4Kj051EORCBSQMycw==",
   "_location": "/nocache",
   "_phantomChildren": {},
   "_requested": {
     "type": "range",
     "registry": true,
-    "raw": "nocache@^2.1.0",
+    "raw": "nocache@^3.0.1",
     "name": "nocache",
     "escapedName": "nocache",
-    "rawSpec": "^2.1.0",
+    "rawSpec": "^3.0.1",
     "saveSpec": null,
-    "fetchSpec": "^2.1.0"
+    "fetchSpec": "^3.0.1"
   },
   "_requiredBy": [
     "/@react-native-community/cli-server-api"
   ],
-  "_resolved": "https://registry.npmjs.org/nocache/-/nocache-2.1.0.tgz",
-  "_shasum": "120c9ffec43b5729b1d5de88cd71aa75a0ba491f",
-  "_spec": "nocache@^2.1.0",
+  "_resolved": "https://registry.npmjs.org/nocache/-/nocache-3.0.4.tgz",
+  "_shasum": "5b37a56ec6e09fc7d401dceaed2eab40c8bfdf79",
+  "_spec": "nocache@^3.0.1",
   "_where": "C:\\Users\\betsa\\ClientsideAPP\\AppReactWebCours\\node_modules\\@react-native-community\\cli-server-api",
   "author": {
     "name": "Adam Baldwin",
@@ -42,33 +42,16 @@
   "deprecated": false,
   "description": "Middleware to destroy caching",
   "devDependencies": {
-    "@types/connect": "^3.4.32",
-    "@types/jest": "^24.0.12",
-    "@types/supertest": "^2.0.7",
-    "@typescript-eslint/eslint-plugin": "^1.7.0",
-    "@typescript-eslint/parser": "^1.7.0",
-    "connect": "^3.6.6",
-    "eslint": "^5.16.0",
-    "eslint-config-helmet": "^0.2.0",
-    "jest": "^24.8.0",
-    "supertest": "^4.0.2",
-    "ts-jest": "^24.0.2",
-    "typescript": "^3.4.5"
+    "connect": "^3.7.0",
+    "eslint": "^8.16.0",
+    "prettier": "^2.6.2",
+    "supertest": "^6.2.3"
   },
   "engines": {
-    "node": ">=4.0.0"
+    "node": ">=12.0.0"
   },
-  "files": [
-    "CHANGELOG.md",
-    "LICENSE",
-    "README.md",
-    "dist/index.js",
-    "dist/index.d.ts"
-  ],
-  "homepage": "https://helmetjs.github.io/docs/nocache/",
+  "homepage": "https://github.com/helmetjs/nocache",
   "keywords": [
-    "helmet",
-    "security",
     "express",
     "connect",
     "nocache",
@@ -76,20 +59,20 @@
     "cache"
   ],
   "license": "MIT",
-  "main": "./dist/index.js",
+  "main": "index.js",
   "name": "nocache",
   "repository": {
     "type": "git",
     "url": "git://github.com/helmetjs/nocache.git"
   },
   "scripts": {
-    "build": "npm run clean && tsc",
-    "clean": "rm -rf dist",
-    "lint": "eslint --fix '**/*.ts'",
-    "prepublishOnly": "npm run build",
+    "format": "prettier --write .",
+    "lint": "npm run lint:eslint && npm run lint:prettier",
+    "lint:eslint": "eslint .",
+    "lint:prettier": "prettier --check .",
     "pretest": "npm run lint",
-    "test": "jest --config test/jest-config.json"
+    "test": "node test"
   },
-  "typings": "./dist/index.d.ts",
-  "version": "2.1.0"
+  "typings": "index.d.ts",
+  "version": "3.0.4"
 }
