@@ -1,26 +1,27 @@
 {
-  "_from": "qs@6.7.0",
-  "_id": "qs@6.7.0",
+  "_from": "qs@6.11.0",
+  "_id": "qs@6.11.0",
   "_inBundle": false,
-  "_integrity": "sha512-VCdBRNFTX1fyE7Nb6FYoURo/SPe62QCaAyzJvUjwRaIsc+NePBEniHlvxFmmX56+HZphIGtV0XeCirBtpDrTyQ==",
+  "_integrity": "sha512-MvjoMCJwEarSbUYk5O+nmoSzSutSsTwF85zcHPQ9OrlFoZOYIjaqBAJIqIXjptyD5vThxGq52Xu/MaJzRkIk4Q==",
   "_location": "/qs",
   "_phantomChildren": {},
   "_requested": {
     "type": "version",
     "registry": true,
-    "raw": "qs@6.7.0",
+    "raw": "qs@6.11.0",
     "name": "qs",
     "escapedName": "qs",
-    "rawSpec": "6.7.0",
+    "rawSpec": "6.11.0",
     "saveSpec": null,
-    "fetchSpec": "6.7.0"
+    "fetchSpec": "6.11.0"
   },
   "_requiredBy": [
-    "/body-parser"
+    "/body-parser",
+    "/expo-linking"
   ],
-  "_resolved": "https://registry.npmjs.org/qs/-/qs-6.7.0.tgz",
-  "_shasum": "41dc1a015e3d581f1621776be31afb2876a9b1bc",
-  "_spec": "qs@6.7.0",
+  "_resolved": "https://registry.npmjs.org/qs/-/qs-6.11.0.tgz",
+  "_shasum": "fd0d963446f7a65e1367e01abd85429453f0c37a",
+  "_spec": "qs@6.11.0",
   "_where": "C:\\Users\\betsa\\ClientsideAPP\\AppReactWebCours\\node_modules\\body-parser",
   "bugs": {
     "url": "https://github.com/ljharb/qs/issues"
@@ -33,27 +34,36 @@
       "url": "http://ljharb.codes"
     }
   ],
-  "dependencies": {},
+  "dependencies": {
+    "side-channel": "^1.0.4"
+  },
   "deprecated": false,
   "description": "A querystring parser that supports nesting and arrays, with a depth limit",
   "devDependencies": {
-    "@ljharb/eslint-config": "^13.1.1",
-    "browserify": "^16.2.3",
-    "covert": "^1.1.1",
-    "editorconfig-tools": "^0.1.1",
-    "eslint": "^5.15.3",
-    "evalmd": "^0.0.17",
+    "@ljharb/eslint-config": "^21.0.0",
+    "aud": "^2.0.0",
+    "browserify": "^16.5.2",
+    "eclint": "^2.8.1",
+    "eslint": "=8.8.0",
+    "evalmd": "^0.0.19",
     "for-each": "^0.3.3",
-    "iconv-lite": "^0.4.24",
-    "mkdirp": "^0.5.1",
-    "object-inspect": "^1.6.0",
+    "has-symbols": "^1.0.3",
+    "iconv-lite": "^0.5.1",
+    "in-publish": "^2.0.1",
+    "mkdirp": "^0.5.5",
+    "npmignore": "^0.3.0",
+    "nyc": "^10.3.2",
+    "object-inspect": "^1.12.2",
     "qs-iconv": "^1.0.4",
-    "safe-publish-latest": "^1.1.2",
+    "safe-publish-latest": "^2.0.0",
     "safer-buffer": "^2.1.2",
-    "tape": "^4.10.1"
+    "tape": "^5.5.3"
   },
   "engines": {
     "node": ">=0.6"
+  },
+  "funding": {
+    "url": "https://github.com/sponsors/ljharb"
   },
   "homepage": "https://github.com/ljharb/qs",
   "keywords": [
@@ -67,20 +77,30 @@
   "license": "BSD-3-Clause",
   "main": "lib/index.js",
   "name": "qs",
+  "publishConfig": {
+    "ignore": [
+      "!dist/*",
+      "bower.json",
+      "component.json",
+      ".github/workflows"
+    ]
+  },
   "repository": {
     "type": "git",
     "url": "git+https://github.com/ljharb/qs.git"
   },
   "scripts": {
-    "coverage": "covert test",
     "dist": "mkdirp dist && browserify --standalone Qs lib/index.js > dist/qs.js",
-    "lint": "eslint lib/*.js test/*.js",
-    "postlint": "editorconfig-tools check * lib/* test/*",
-    "prepublish": "safe-publish-latest && npm run dist",
+    "lint": "eslint --ext=js,mjs .",
+    "postlint": "eclint check $(git ls-files | xargs find 2> /dev/null | grep -vE 'node_modules|\\.git' | grep -v dist/)",
+    "posttest": "aud --production",
+    "prepack": "npmignore --auto --commentLines=autogenerated",
+    "prepublish": "not-in-publish || npm run prepublishOnly",
+    "prepublishOnly": "safe-publish-latest && npm run dist",
     "pretest": "npm run --silent readme && npm run --silent lint",
     "readme": "evalmd README.md",
-    "test": "npm run --silent coverage",
-    "tests-only": "node test"
+    "test": "npm run tests-only",
+    "tests-only": "nyc tape 'test/**/*.js'"
   },
-  "version": "6.7.0"
+  "version": "6.11.0"
 }
