@@ -1,26 +1,26 @@
 {
-  "_from": "raw-body@2.4.0",
-  "_id": "raw-body@2.4.0",
+  "_from": "raw-body@2.5.2",
+  "_id": "raw-body@2.5.2",
   "_inBundle": false,
-  "_integrity": "sha512-4Oz8DUIwdvoa5qMJelxipzi/iJIi40O5cGV1wNYp5hvZP8ZN0T+jiNkL0QepXs+EsQ9XJ8ipEDoiH70ySUJP3Q==",
+  "_integrity": "sha512-8zGqypfENjCIqGhgXToC8aB2r7YrBX+AQAfIPs/Mlk+BtPTztOvTS01NRW/3Eh60J+a48lt8qsCzirQ6loCVfA==",
   "_location": "/raw-body",
   "_phantomChildren": {},
   "_requested": {
     "type": "version",
     "registry": true,
-    "raw": "raw-body@2.4.0",
+    "raw": "raw-body@2.5.2",
     "name": "raw-body",
     "escapedName": "raw-body",
-    "rawSpec": "2.4.0",
+    "rawSpec": "2.5.2",
     "saveSpec": null,
-    "fetchSpec": "2.4.0"
+    "fetchSpec": "2.5.2"
   },
   "_requiredBy": [
     "/body-parser"
   ],
-  "_resolved": "https://registry.npmjs.org/raw-body/-/raw-body-2.4.0.tgz",
-  "_shasum": "a1ce6fb9c9bc356ca52e89256ab59059e13d0332",
-  "_spec": "raw-body@2.4.0",
+  "_resolved": "https://registry.npmjs.org/raw-body/-/raw-body-2.5.2.tgz",
+  "_shasum": "99febd83b90e08975087e8f1f9419a149366b68a",
+  "_spec": "raw-body@2.5.2",
   "_where": "C:\\Users\\betsa\\ClientsideAPP\\AppReactWebCours\\node_modules\\body-parser",
   "author": {
     "name": "Jonathan Ong",
@@ -42,26 +42,26 @@
     }
   ],
   "dependencies": {
-    "bytes": "3.1.0",
-    "http-errors": "1.7.2",
+    "bytes": "3.1.2",
+    "http-errors": "2.0.0",
     "iconv-lite": "0.4.24",
     "unpipe": "1.0.0"
   },
   "deprecated": false,
   "description": "Get and validate the raw body of a readable stream.",
   "devDependencies": {
-    "bluebird": "3.5.4",
-    "eslint": "5.16.0",
-    "eslint-config-standard": "12.0.0",
-    "eslint-plugin-import": "2.16.0",
-    "eslint-plugin-markdown": "1.0.0",
-    "eslint-plugin-node": "8.0.1",
-    "eslint-plugin-promise": "4.1.1",
-    "eslint-plugin-standard": "4.0.0",
-    "istanbul": "0.4.5",
-    "mocha": "6.1.3",
-    "readable-stream": "2.3.6",
-    "safe-buffer": "5.1.2"
+    "bluebird": "3.7.2",
+    "eslint": "8.34.0",
+    "eslint-config-standard": "15.0.1",
+    "eslint-plugin-import": "2.27.5",
+    "eslint-plugin-markdown": "3.0.0",
+    "eslint-plugin-node": "11.1.0",
+    "eslint-plugin-promise": "6.1.1",
+    "eslint-plugin-standard": "4.1.0",
+    "mocha": "10.2.0",
+    "nyc": "15.1.0",
+    "readable-stream": "2.3.7",
+    "safe-buffer": "5.2.1"
   },
   "engines": {
     "node": ">= 0.8"
@@ -70,6 +70,7 @@
     "HISTORY.md",
     "LICENSE",
     "README.md",
+    "SECURITY.md",
     "index.d.ts",
     "index.js"
   ],
@@ -81,10 +82,10 @@
     "url": "git+https://github.com/stream-utils/raw-body.git"
   },
   "scripts": {
-    "lint": "eslint --plugin markdown --ext js,md .",
+    "lint": "eslint .",
     "test": "mocha --trace-deprecation --reporter spec --bail --check-leaks test/",
-    "test-cov": "istanbul cover node_modules/mocha/bin/_mocha -- --trace-deprecation --reporter dot --check-leaks test/",
-    "test-travis": "istanbul cover node_modules/mocha/bin/_mocha --report lcovonly -- --trace-deprecation --reporter spec --check-leaks test/"
+    "test-ci": "nyc --reporter=lcovonly --reporter=text npm test",
+    "test-cov": "nyc --reporter=html --reporter=text npm test"
   },
-  "version": "2.4.0"
+  "version": "2.5.2"
 }
