@@ -1,30 +1,29 @@
 {
-  "_from": "react-dom@17.0.2",
-  "_id": "react-dom@17.0.2",
+  "_from": "react-dom@18.1.0",
+  "_id": "react-dom@18.1.0",
   "_inBundle": false,
-  "_integrity": "sha512-s4h96KtLDUQlsENhMn1ar8t2bEa+q/YAtj8pPPdIjPDGBDIVNsrD9aXNWqspUe6AzKCIG0C1HZZLqLV7qpOBGA==",
+  "_integrity": "sha512-fU1Txz7Budmvamp7bshe4Zi32d0ll7ect+ccxNu9FlObT605GOEB8BfO4tmRJ39R5Zj831VCpvQ05QPBW5yb+w==",
   "_location": "/react-dom",
   "_phantomChildren": {},
   "_requested": {
     "type": "version",
     "registry": true,
-    "raw": "react-dom@17.0.2",
+    "raw": "react-dom@18.1.0",
     "name": "react-dom",
     "escapedName": "react-dom",
-    "rawSpec": "17.0.2",
+    "rawSpec": "18.1.0",
     "saveSpec": null,
-    "fetchSpec": "17.0.2"
+    "fetchSpec": "18.1.0"
   },
   "_requiredBy": [
     "/"
   ],
-  "_resolved": "https://registry.npmjs.org/react-dom/-/react-dom-17.0.2.tgz",
-  "_shasum": "ecffb6845e3ad8dbfcdc498f0d0a939736502c23",
-  "_spec": "react-dom@17.0.2",
+  "_resolved": "https://registry.npmjs.org/react-dom/-/react-dom-18.1.0.tgz",
+  "_shasum": "7f6dd84b706408adde05e1df575b3a024d7e8a2f",
+  "_spec": "react-dom@18.1.0",
   "_where": "C:\\Users\\betsa\\ClientsideAPP\\AppReactWebCours",
   "browser": {
-    "./server.js": "./server.browser.js",
-    "./unstable-fizz.js": "./unstable-fizz.browser.js"
+    "./server.js": "./server.browser.js"
   },
   "browserify": {
     "transform": [
@@ -37,16 +36,30 @@
   "bundleDependencies": false,
   "dependencies": {
     "loose-envify": "^1.1.0",
-    "object-assign": "^4.1.1",
-    "scheduler": "^0.20.2"
+    "scheduler": "^0.22.0"
   },
   "deprecated": false,
   "description": "React package for working with the DOM.",
+  "exports": {
+    ".": "./index.js",
+    "./client": "./client.js",
+    "./server": {
+      "deno": "./server.browser.js",
+      "worker": "./server.browser.js",
+      "browser": "./server.browser.js",
+      "default": "./server.node.js"
+    },
+    "./server.browser": "./server.browser.js",
+    "./server.node": "./server.node.js",
+    "./profiling": "./profiling.js",
+    "./test-utils": "./test-utils.js",
+    "./package.json": "./package.json"
+  },
   "files": [
     "LICENSE",
     "README.md",
-    "build-info.json",
     "index.js",
+    "client.js",
     "profiling.js",
     "server.js",
     "server.browser.js",
@@ -63,7 +76,7 @@
   "main": "index.js",
   "name": "react-dom",
   "peerDependencies": {
-    "react": "17.0.2"
+    "react": "^18.1.0"
   },
   "repository": {
     "type": "git",
@@ -73,5 +86,5 @@
   "scripts": {
     "start": "node server.js"
   },
-  "version": "17.0.2"
+  "version": "18.1.0"
 }
