@@ -54,28 +53,33 @@
   "deprecated": false,
   "description": "A flexible way to handle safe area, also works on Android and web.",
   "devDependencies": {
-    "@react-native-community/eslint-config": "^3.0.1",
-    "@react-native/eslint-plugin-specs": "^0.0.3",
-    "@types/jest": "^27.4.0",
-    "@types/react": "^17.0.39",
-    "@types/react-dom": "^17.0.11",
-    "@types/react-native": "^0.66.16",
-    "@types/react-test-renderer": "^17.0.1",
-    "@typescript-eslint/eslint-plugin": "^5.12.0",
-    "@typescript-eslint/parser": "^5.12.0",
+    "@commitlint/config-conventional": "^17.0.3",
+    "@react-native-community/eslint-config": "^3.1.0",
+    "@react-native/eslint-plugin-specs": "^0.71.0",
+    "@release-it/conventional-changelog": "^5.1.0",
+    "@types/jest": "^28.1.7",
+    "@types/react": "^18.0.17",
+    "@types/react-dom": "^18.0.6",
+    "@types/react-native": "^0.69.5",
+    "@types/react-test-renderer": "^18.0.0",
+    "@typescript-eslint/eslint-plugin": "^5.33.1",
+    "@typescript-eslint/parser": "^5.33.1",
     "babel-plugin-module-resolver": "^4.1.0",
-    "eslint": "^7.32.0",
-    "eslint-config-prettier": "^8.3.0",
-    "eslint-plugin-prettier": "^4.0.0",
-    "jest": "^27.5.1",
-    "metro-react-native-babel-preset": "^0.68.0",
-    "prettier": "^2.5.1",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-native": "0.0.0-20220213-2008-7cece3423",
-    "react-native-builder-bob": "^0.18.2",
-    "react-test-renderer": "^17.0.2",
-    "typescript": "^4.5.5"
+    "commitlint": "^17.0.3",
+    "eslint": "^8.22.0",
+    "eslint-config-prettier": "^8.5.0",
+    "eslint-plugin-prettier": "^4.2.1",
+    "husky": "^8.0.1",
+    "jest": "^28.1.3",
+    "metro-react-native-babel-preset": "^0.72.1",
+    "prettier": "^2.7.1",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-native": "0.69.4",
+    "react-native-builder-bob": "^0.18.3",
+    "react-test-renderer": "^18.2.0",
+    "release-it": "^15.3.0",
+    "typescript": "^4.7.4"
   },
   "files": [
     "src",
