@@ -31,58 +31,39 @@
   },
   "bundleDependencies": false,
   "dependencies": {
-    "@react-native-community/cli-debugger-ui": "^7.0.3",
-    "@react-native-community/cli-hermes": "^6.3.1",
-    "@react-native-community/cli-plugin-metro": "^7.0.4",
-    "@react-native-community/cli-server-api": "^7.0.4",
-    "@react-native-community/cli-tools": "^6.2.1",
-    "@react-native-community/cli-types": "^6.0.0",
-    "appdirsjs": "^1.2.4",
+    "@react-native-community/cli-clean": "^9.2.1",
+    "@react-native-community/cli-config": "^9.2.1",
+    "@react-native-community/cli-debugger-ui": "^9.0.0",
+    "@react-native-community/cli-doctor": "^9.2.1",
+    "@react-native-community/cli-hermes": "^9.2.1",
+    "@react-native-community/cli-plugin-metro": "^9.2.1",
+    "@react-native-community/cli-server-api": "^9.2.1",
+    "@react-native-community/cli-tools": "^9.2.1",
+    "@react-native-community/cli-types": "^9.1.0",
     "chalk": "^4.1.2",
-    "command-exists": "^1.2.8",
-    "commander": "^2.19.0",
-    "cosmiconfig": "^5.1.0",
-    "deepmerge": "^3.2.0",
-    "envinfo": "^7.7.2",
+    "commander": "^9.4.0",
     "execa": "^1.0.0",
     "find-up": "^4.1.0",
     "fs-extra": "^8.1.0",
-    "glob": "^7.1.3",
     "graceful-fs": "^4.1.3",
-    "joi": "^17.2.1",
-    "leven": "^3.1.0",
-    "lodash": "^4.17.15",
-    "minimist": "^1.2.0",
-    "node-stream-zip": "^1.9.1",
-    "ora": "^3.4.0",
-    "pretty-format": "^26.6.2",
     "prompts": "^2.4.0",
-    "semver": "^6.3.0",
-    "serve-static": "^1.13.1",
-    "strip-ansi": "^5.2.0",
-    "sudo-prompt": "^9.0.0",
-    "wcwidth": "^1.0.1"
+    "semver": "^6.3.0"
   },
   "deprecated": false,
   "description": "React Native CLI",
   "devDependencies": {
-    "@types/command-exists": "^1.2.0",
-    "@types/cosmiconfig": "^5.0.3",
     "@types/fs-extra": "^8.1.0",
-    "@types/glob": "^7.1.1",
     "@types/graceful-fs": "^4.1.3",
     "@types/hapi__joi": "^17.1.6",
-    "@types/lodash": "^4.14.149",
-    "@types/minimist": "^1.2.0",
     "@types/prompts": "^2.0.9",
     "@types/semver": "^6.0.2",
-    "@types/wcwidth": "^1.0.0",
+    "deepmerge": "^3.2.0",
     "slash": "^3.0.0",
     "snapshot-diff": "^0.7.0"
   },
   "engineStrict": true,
   "engines": {
-    "node": ">=12"
+    "node": ">=14"
   },
   "files": [
     "build",
