@@ -30,41 +30,83 @@
     "url": "https://github.com/tj/commander.js/issues"
   },
   "bundleDependencies": false,
-  "dependencies": {},
   "deprecated": false,
   "description": "the complete solution for node.js command-line programs",
   "devDependencies": {
-    "@types/node": "^12.7.8",
-    "eslint": "^6.4.0",
-    "should": "^13.2.3",
-    "sinon": "^7.5.0",
-    "standard": "^14.3.1",
-    "ts-node": "^8.4.1",
-    "typescript": "^3.6.3"
+    "@types/jest": "^28.1.4",
+    "@types/node": "^16.11.15",
+    "@typescript-eslint/eslint-plugin": "^5.30.6",
+    "@typescript-eslint/parser": "^5.30.6",
+    "eslint": "^8.19.0",
+    "eslint-config-standard": "^17.0.0",
+    "eslint-config-standard-with-typescript": "^22.0.0",
+    "eslint-plugin-import": "^2.25.3",
+    "eslint-plugin-jest": "^26.5.3",
+    "eslint-plugin-n": "^15.2.4",
+    "eslint-plugin-promise": "^6.0.0",
+    "jest": "^28.1.2",
+    "ts-jest": "^28.0.5",
+    "tsd": "^0.22.0",
+    "typescript": "^4.7.4"
+  },
+  "engines": {
+    "node": "^12.20.0 || >=14"
+  },
+  "exports": {
+    ".": {
+      "types": "./typings/index.d.ts",
+      "require": "./index.js",
+      "import": "./esm.mjs"
+    },
+    "./esm.mjs": "./esm.mjs"
   },
   "files": [
     "index.js",
-    "typings/index.d.ts"
+    "lib/*.js",
+    "esm.mjs",
+    "typings/index.d.ts",
+    "package-support.json"
   ],
   "homepage": "https://github.com/tj/commander.js#readme",
+  "jest": {
+    "testEnvironment": "node",
+    "collectCoverage": true,
+    "transform": {
+      "^.+\\.tsx?$": "ts-jest"
+    },
+    "testPathIgnorePatterns": [
+      "/node_modules/"
+    ]
+  },
   "keywords": [
     "commander",
     "command",
     "option",
-    "parser"
+    "parser",
+    "cli",
+    "argument",
+    "args",
+    "argv"
   ],
   "license": "MIT",
-  "main": "index",
+  "main": "./index.js",
   "name": "commander",
   "repository": {
     "type": "git",
     "url": "git+https://github.com/tj/commander.js.git"
   },
   "scripts": {
-    "lint": "eslint index.js",
-    "test": "node test/run.js && npm run test-typings",
-    "test-typings": "tsc -p tsconfig.json"
+    "lint": "npm run lint:javascript && npm run lint:typescript",
+    "lint:javascript": "eslint index.js esm.mjs \"lib/*.js\" \"tests/**/*.js\"",
+    "lint:typescript": "eslint typings/*.ts tests/*.ts",
+    "test": "jest && npm run test-typings",
+    "test-all": "npm run test && npm run lint && npm run typescript-checkJS && npm run test-esm",
+    "test-esm": "node --experimental-modules ./tests/esm-imports-test.mjs",
+    "test-typings": "tsd",
+    "typescript-checkJS": "tsc --allowJS --checkJS index.js lib/*.js --noEmit"
   },
-  "typings": "typings/index.d.ts",
-  "version": "2.20.3"
+  "support": true,
+  "type": "commonjs",
+  "types": "typings/index.d.ts",
+  "version": "9.5.0"
 }
